requirements.txt
Flask>=2.0
python-docx>=0.8.11
docx2pdf>=0.1.8
img2pdf>=0.4.0
pypandoc>=1.11
Pillow>=9.0.0
pywin32>=227 